+++ Solution/RiskOfChaos/Patches/AttackHooks/BlastAttackHookManager.cs
@@ -1,0 +1,20 @@
+﻿using RiskOfChaos.Utilities;
+using RoR2;
+
+namespace RiskOfChaos.Patches.AttackHooks
+{
+    class BlastAttackHookManager : AttackHookManager
+    {
+        readonly BlastAttack _blastAttack;
+
+        public BlastAttackHookManager(BlastAttack blastAttack)
+        {
+            _blastAttack = AttackUtils.Clone(blastAttack);
+        }
+
+        protected override void fireAttackCopy()
+        {
+            AttackUtils.Clone(_blastAttack).Fire();
+        }
+    }
+}